pyproject.toml
[project]
name = "bc-sql-aula15"
version = "0.1.0"
description = ""
authors = [
    {name = "Renan Domingos",email = "rna.domingos@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "apache-airflow (>=3.0.2,<4.0.0)",
    "apache-airflow-providers-dbt-cloud (>=4.4.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
